refactor: remove all clippy suppressions by fixing root causes
- Delete [lints.clippy] block from Cargo.toml entirely
- Remove postcode field from CameraProperties (was dead code; serde
  ignores unknown fields automatically)
- Change refresh_interval_secs u64→u32 so i64::from() is safe;
  change camera_aspect_ratio [u32;2]→[u16;2] so as f32 is lossless
- Rewrite egui sliders to use native field types (u32, usize, u8, f32)
  directly in Slider::new — eliminates all f64 round-trip casts
- Rewrite if-let-else { None } patterns as match in app.rs
- Extract section helpers (show_districts, show_cameras, show_display,
  show_disk_save, ordered_camera_ids) from settings::show to fix
  too_many_lines; extract render_tile from grid::show likewise
- Replace remaining unavoidable suppresses with narrow per-line
  #[allow] on two geometry lines in grid.rs (cast_precision_loss,
  suboptimal_flops) and two test assertions in config.rs (float_cmp)
- Delete .vscode/settings.json evenBetterToml workaround (no longer
  needed now that [lints.clippy] is gone)
- Update CLAUDE.md and CONTRIBUTING.md lint policy docs

+++ src/config.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeSet;
 use std::path::PathBuf;
 
 const APP_NAME: &str = "traffic-camera-viewer";
-const NATIVE_CAMERA_IMAGE_SIZE: [u32; 2] = [320, 256];
+const NATIVE_CAMERA_IMAGE_SIZE: [u16; 2] = [320, 256];
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(default)]
@@ -11,7 +11,7 @@ pub struct AppConfig {
     /// Districts whose cameras are shown.
     pub selected_districts: BTreeSet<String>,
     /// Auto-refresh period in seconds (10–300).
-    pub refresh_interval_secs: u64,
+    pub refresh_interval_secs: u32,
     /// Number of columns in the image grid (1–6).
     pub column_count: usize,
     /// Save fetched images to disk.
@@ -29,7 +29,7 @@ pub struct AppConfig {
     /// Camera title text color as RGB.
     pub camera_title_rgb: [u8; 3],
     /// Camera image tile aspect ratio as width:height.
-    pub camera_aspect_ratio: [u32; 2],
+    pub camera_aspect_ratio: [u16; 2],
     /// Grid/background color as RGB (does not affect status bar panel).
     pub app_background_rgb: [u8; 3],
     /// Camera IDs that have been manually hidden by the user.
